changed configuration to optionally generate more warnings
* renamed `definitions` to `cu0_definitions` in `CMakeLists.txt`
* added optional `CU0_ENABLE_MORE_WARNINGS` flag
* changed `min_stack_size` type in `checks/check_cu0_strand.cc`
* introduced `[[maybe_unused]]` in `examples/example_cu0_strand_create.cc`
* removed a comparison in `measurements/measurement_cu0_error_return.cc`

+++ checks/check_cu0_strand.cc
@@ -447,8 +447,9 @@ be checked
 #if __has_include(<stdlib.h>) && __has_include(<unistd.h>)
       const auto page_size = sysconf(_SC_PAGE_SIZE);
       assert(page_size > 0);
-      const auto min_stack_size =
-          (PTHREAD_STACK_MIN / page_size + 1) * page_size;
+      const auto min_stack_size = static_cast<
+          unsigned long
+      >(PTHREAD_STACK_MIN / page_size + 1) * page_size;
     {
       const auto allocate_stack_variant =
           strand_for_stack_size.allocate_stack<
